[JBIDE-21352] server adapter wiz: preselecting project select in project views
We now preselect the (1st) project that's selected in (using this order of prescedence)
* package explorer
* project explorer
* resource navigator
